+++ SoftRasterLearning/main.cpp
@@ -29,7 +29,7 @@ int WINAPI WinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPreInstance, _I
 	wnd::DC_WND wnd = wnd::DC_WND{ hInstance }.WndClassName(L"dc_wnd_cls").WndName(L"dc_wnd_wnd").Size(800, 600).AddWndStyle(~WS_MAXIMIZEBOX).Init().Move();
 
 	sr::Context ctx;
-	ctx.Viewport(800, 600, { 0.4f, 0.6f, 0.2f, 1.f });
+	ctx.Viewport(800, 600, { 0.4, 0.6, 0.2, 1. });
 
 	Material_Model m;
 
@@ -44,21 +44,21 @@ int WINAPI WinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPreInstance, _I
 		{{-0.5,-0.5,0,1},{1,0,0,1}}
 	};
 
-	float pi = 3.14159265358979f;
+	double pi = 3.14159265358979;
 	
-	m.mat = sr::Mat::Translate(0.3f, 0, 0) * sr::Mat::Rotate(0, 0, -40.f / 180.f * pi) * sr::Mat::Scale(1.0f, 1.0f, 1) * m.mat;
+	m.mat = sr::Mat::Translate(0.3f, 0, 0) * sr::Mat::Rotate(0, 0, -40. / 180. * pi) * sr::Mat::Scale(1.0f, 1.0f, 1) * m.mat;
 	
 	renderer.DrawTriangles(rect,6);
 
 	ctx.CopyToScreen(wnd.getFrameBufferView());
 	wnd.drawBuffer();
 
-	float time = 0;
+	double time = 0;
 	while (!wnd.app_should_close())
 	{
 		wnd.PeekMsg();
-		ctx.Clear({ 0.4f, 0.6f, 0.2f, 1.f });
-		m.mat = sr::Mat::Translate(sin(time)-0.4f, 0, 100) * sr::Mat::Rotate(0, 0, time+=0.1f / 180 * pi) * sr::Mat::Scale(1.f, 1.f, 1);// *m.mat;
+		ctx.Clear({ 0.4, 0.6, 0.2, 1. });
+		m.mat = sr::Mat::Translate(sin(time)-0.4, 0, 100) * sr::Mat::Rotate(0, 0, time+=0.1 / 180 * pi) * sr::Mat::Scale(1., 1., 1);// *m.mat;
 		renderer.DrawTriangles(rect, 6);
 		ctx.CopyToScreen(wnd.getFrameBufferView());
 		wnd.drawBuffer();
